found panic

+++ src/structs/skill_progression.rs
@@ -4,5 +4,5 @@ use utoipa::ToSchema;
 #[derive(Serialize, ToSchema)]
 pub struct SkillProgression {
     pub skill_name: String,
-    pub progression: f32,
+    pub progression: f64,
 }
